api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Thalida Noel <shoplist@thalida.com>"]
license = "GNU AFFERO GENERAL PUBLIC LICENSE"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
graphene-django = "^3.1.2"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
django-unfold = "^0.6.8"
django-cors-headers = "^4.1.0"
django-filter = "^23.2"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
django-oauth-toolkit = "^2.3.0"
djangorestframework = "^3.14.0"
drf-social-oauth2 = "^2.1.3"
drf-spectacular = "^0.26.3"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
django-graphql-jwt = "^0.3.4"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.3"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
pylint-django = "^2.5.3"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
